Umsetzung mit BiFunction

+++ Operator.java
@@ -8,5 +8,5 @@
 
 public interface Operator
 {
-    double calculate(double a, double b);
+    int calculate(int a, int b);
 }
